Move plot related actions into it's own file
